Flood Disaster Early Warning Connected to NodeMCU-ESP8266 WiFi Board
This project was used to measure the water inflow rate and water level flowing into a container. The container is installed with water flow and level sensor and wireless sensor network; ESP8266 was used to send the data to Thingspeak channel with the API key and WiFi API name included in the programming code. The data from the sensor were then processed and used to train a 2-Class Neural Network using Azure Machine Learning.