Azure ML Workspace Python Script
This is a code of the Azure Machine Learning Python Script derived from the Generate Data Access Code of the convert to CSV node in the Azure ML workspace. This code includes the workspace access token, which provides full access to the workspace. Therefore, the area marked with asterisks is personal access tokens derived from the personal workspace.